requirements.txt
appdirs==1.4.4
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudflare==2.8.15
colorama==0.4.4
idna==3.3
jsonlines==2.0.0
loguru==0.5.3
oauthlib==3.1.1
pycodestyle==2.8.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
tweepy==4.4.0
urllib3==1.26.7
websocket-client==1.2.1
win32-setctime==1.0.3
youtube-dl==2021.6.6
